feat: install express and create registration and login route
