--- @types/alltypes.d.ts
@@ -1,3 +1,0 @@
-declare module 'react-materialize';
-declare module 'react-router';
-declare module 'flux';
