+++ src/components/cards/country.tsx
@@ -1,0 +1,11 @@
+export default function Country ({match}) {
+  return (
+    <div className = "country-page">
+      <div className="widget-wrapper">
+        <div className="local-time"></div>
+        <div className="Local-wallet"></div>
+        <div className="local-weather"></div>
+      </div>
+    </div>
+  )
+}
