+++ src/countries.js
@@ -1,0 +1,58 @@
+const slides = [
+  {
+    id: 'switzerland',
+    title: 'Switzerland',
+    description: 'Heart-racing Alpine pursuits and urban culture.',
+    image: 'https://img3.goodfon.ru/original/1920x1080/4/43/zermatt-valley-switzerland.jpg',
+
+  },
+  {
+    id: 'indonesia',
+    title: 'Indonesia',
+    description: 'Explore The magic of Indonesia',
+    image: 'https://images.wallpaperscraft.ru/image/indoneziya_ostrov_yava_vulkan_bromo_holmy_vysota_goluboe_rozovoe_nebo_oblaka_90613_1920x1080.jpg',
+  },
+  {
+    id: 'newZealand',
+    title: 'New Zealand',
+    description: 'Adventure is never far away',
+    image: 'https://cdn.hipwallpaper.com/i/35/72/fcT04X.jpg',
+
+  },
+  {
+    id: 'iceland',
+    title: 'Iceland',
+    description: 'Let your dreams come true',
+    image: 'https://img1.goodfon.ru/original/1920x1080/e/ff/more-gory-doma-islandiya.jpg',
+
+  },
+  {
+    id: 'thailand',
+    title: 'Thailand',
+    description: 'Sun and beaches all year around',
+    image: 'https://wallpaperaccess.com/full/150582.jpg',
+
+  },
+  {
+    id: 'india',
+    title: 'India',
+    description: 'Home of tigers elephants and monkeys ',
+    image: 'https://media.cntraveller.in/wp-content/uploads/2020/03/Kerala-Coronavirus-measures_Periyar.jpg',
+
+  },
+  {
+    id: 'germany',
+    title: 'Germany',
+    description: 'Fairy tale about medieval fortresses with moats',
+    image: 'https://wallpaperaccess.com/full/138772.jpg',
+
+  },
+  {
+    id: 'france',
+    title: 'France',
+    description: 'The mountains are calling',
+    image: 'https://s1.1zoom.me/b5050/562/France_Sunrises_and_sunsets_Mountains_Stones_566654_1920x1080.jpg',
+  }
+]
+
+export default slides;
